+++ _BIOSYNTH_PROJECTS/Biosynth_project_loader/src/configuration.h
@@ -18,7 +18,7 @@
 
 namespace configuration{
     
-    const int board_id{1};
+    const int board_id{4};
  
     const float volume{0.6}; //Initial volume of the board
